+++ shopping_list/locationadmin.php
@@ -67,7 +67,9 @@ while($row=$result->fetch_assoc()) {
 			var myID = $(this).closest("li").attr("id");
 			var myTitle = $(this).closest("li").text();
 			if (confirm('Are you sure you want to delete '+myTitle+'?')) {
-				DeleteLocationRecord(myID.replace("item-", ""));
+				if(confirm('This will delete all instances of '+myTitle+' and cannot be undone.  Are you sure you want to proceed?')) {
+					DeleteLocationRecord(myID.replace("item-", ""));
+				}
 			}
 		})
 		function DeleteLocationRecord(id) {
